Fix osquery cypress tests (#163988)
## Summary

Adjust tests to https://github.com/elastic/kibana/pull/161614
Split tests into smaller files to better utilize parallelization and
increase the stability of tests

(cherry picked from commit fd33ed55fd9bc81d006ca41c85b7bd4117741e80)
