+++ .buildkite/pipelines/pull_request/osquery_cypress.yml
@@ -12,3 +12,13 @@ steps:
           limit: 1
     artifact_paths:
       - "target/kibana-osquery/**/*"
+
+  - command: .buildkite/scripts/steps/functional/osquery_cypress_burn.sh
+    label: 'Osquery Cypress Tests, burning changed specs'
+    agents:
+      queue: n2-4-spot
+    depends_on: build
+    timeout_in_minutes: 50
+    soft_fail: true
+    artifact_paths:
+      - "target/kibana-osquery/**/*"
