+++ x-pack/plugins/osquery/cypress/e2e/roles/alert_test.cy.ts
@@ -97,6 +97,7 @@ describe('Alert Test', () => {
       cy.getBySel('expand-event').first().click();
 
       cy.wait(500);
+      cy.getBySel('securitySolutionDocumentDetailsFlyoutInvestigationGuideButton').click();
       cy.contains('Get processes').click();
     });
