+++ x-pack/plugins/osquery/cypress/tasks/api_fixtures.ts
@@ -229,6 +229,7 @@ export const cleanupCase = (id: string) => {
     method: 'DELETE',
     url: '/api/cases',
     qs: { ids: JSON.stringify([id]) },
+    failOnStatusCode: false,
   });
 };
