+++ x-pack/plugins/osquery/cypress/tasks/live_query.ts
@@ -6,6 +6,7 @@
  */
 
 import { LIVE_QUERY_EDITOR } from '../screens/live_query';
+import { ROLE, login } from './login';
 
 export const DEFAULT_QUERY = 'select * from processes;';
 export const BIG_QUERY = 'select * from processes, users limit 110;';
@@ -100,6 +101,7 @@ export const toggleRuleOffAndOn = (ruleName: string) => {
 };
 
 export const loadRuleAlerts = (ruleName: string) => {
+  login(ROLE.soc_manager);
   cy.visit('/app/security/rules');
   cy.contains(ruleName).click();
   cy.getBySel('alertsTable').within(() => {
