+++ x-pack/plugins/osquery/package.json
@@ -5,8 +5,12 @@
   "private": true,
   "license": "Elastic License 2.0",
   "scripts": {
+    "cypress:burn": "yarn cypress:run --env burn=2 --concurrency=1 --headed",
+    "cypress:changed-specs-only": "yarn cypress:run --changed-specs-only --env burn=2",
     "cypress:open": "node ../security_solution/scripts/start_cypress_parallel open --config-file ../osquery/cypress.config.ts --ftr-config-file ../../../x-pack/test/osquery_cypress/cli_config",
     "cypress:run": "node ../security_solution/scripts/start_cypress_parallel run --config-file ../osquery/cypress.config.ts --ftr-config-file ../../../x-pack/test/osquery_cypress/cli_config --concurrency 1",
-    "nyc": "../../../node_modules/.bin/nyc report --reporter=text-summary"
+    "nyc": "../../../node_modules/.bin/nyc report --reporter=text-summary",
+    "junit:merge": "../../../node_modules/.bin/mochawesome-merge ../../../target/kibana-osquery/cypress/results/mochawesome*.json > ../../../target/kibana-osquery/cypress/results/output.json && ../../../node_modules/.bin/marge ../../../target/kibana-osquery/cypress/results/output.json --reportDir ../../../target/kibana-osquery/cypress/results && yarn junit:transform && mkdir -p ../../../target/junit && cp ../../../target/kibana-osquery/cypress/results/*.xml ../../../target/junit/",
+    "junit:transform": "node ../security_solution/scripts/junit_transformer --pathPattern '../../../target/kibana-osquery/cypress/results/*.xml' --rootDirectory ../../../ --reportName 'Osquery Cypress' --writeInPlace"
   }
 }
