+++ x-pack/plugins/security_solution/scripts/run_cypress/parallel.ts
@@ -105,6 +105,10 @@ export const cli = () => {
           // eslint-disable-next-line no-process-exit
           return process.exit(0);
         }
+
+        // to avoid running too many tests, we limit the number of files to 3
+        // we may extend this in the future
+        files = files.slice(0, 3);
       }
 
       if (!files?.length) {
